added the additonal arguments to the clean function in lima_segmenter module
